LayoutTests/fast/forms/checkValidity-001-expected.txt

This test checks if checkValidity() returns correctly a true (meaning no error) result.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


 lorem ipsum  
PASS v[i].checkValidity() is true
PASS v[i].checkValidity() is true
PASS v[i].checkValidity() is true
PASS v[i].checkValidity() is true
PASS v[i].checkValidity() is true
PASS successfullyParsed is true

TEST COMPLETE
